couple small patches
