+++ app/src/components/shared/constants/timetable.constants.ts
@@ -1,0 +1,10 @@
+export type Term = 'SUMMER' | 'WINTER' | 'SPRING';
+
+export type Faculty = 'CSC' | 'SENG' | 'ECE';
+
+export enum Role {
+  User = 'USER',
+  Admin = 'ADMIN'
+}
+
+export type WeekDay = 'MONDAY' | 'TUESDAY' | 'WEDNESDAY' | 'THURSDAY' | 'FRIDAY' | 'SATURDAY' | 'SUNDAY';
